- Corrected the tests in test_node_suspension.py according to the recent changes. - Extracted some common code from CLI tests to conftest and helper modules.
