+++ sovrin_client/test/cli/conftest.py
@@ -8,7 +8,11 @@ from typing import List
 import pytest
 
 import plenum
+from plenum.common import util
+from plenum.common.constants import ALIAS
+from plenum.common.constants import CLIENT_STACK_SUFFIX
 from plenum.common.exceptions import BlowUp
+from plenum.test import waits
 from stp_core.loop.eventually import eventually
 from stp_core.common.log import getlogger
 from plenum.test.conftest import tdirWithPoolTxns, tdirWithDomainTxns
@@ -29,7 +33,7 @@ from plenum.test.cli.helper import newKeyPair, waitAllNodesStarted, \
 from sovrin_common.config_util import getConfig
 from sovrin_client.test.cli.helper import ensureNodesCreated, getLinkInvitation, \
     getPoolTxnData, newCLI, getCliBuilder, P, prompt_is, addAgent, \
-    getNewNodeVals
+    getNewNodeVals, doSendNodeCmd
 from sovrin_client.test.agent.conftest import faberIsRunning as runningFaber, \
     acmeIsRunning as runningAcme, thriftIsRunning as runningThrift, emptyLooper,\
     faberWallet, acmeWallet, thriftWallet, agentIpAddress, \
@@ -1353,3 +1357,48 @@ def newStewardCli(be, do, poolNodesStarted, trusteeCli,
            expect=connectedToTest)
 
     return cliWithNewStewardName
+
+
+@pytest.fixture(scope="module")
+def newNodeAdded(be, do, poolNodesStarted, philCli, newStewardCli,
+                 connectedToTest, newNodeVals):
+    be(philCli)
+
+    if not philCli._isConnectedToAnyEnv():
+        do('connect test', within=3,
+           expect=connectedToTest)
+
+    be(newStewardCli)
+    doSendNodeCmd(do, newNodeVals)
+    newNodeData = newNodeVals["newNodeData"]
+
+    def checkClientConnected(client):
+        name = newNodeData[ALIAS] + CLIENT_STACK_SUFFIX
+        assert name in client.nodeReg
+
+    def checkNodeConnected(nodes):
+        for node in nodes:
+            name = newNodeData[ALIAS]
+            assert name in node.nodeReg
+
+    timeout = waits.expectedClientToPoolConnectionTimeout(
+        util.getMaxFailures(len(philCli.nodeReg))
+    )
+    newStewardCli.looper.run(eventually(checkClientConnected,
+                                        newStewardCli.activeClient,
+                                        timeout=timeout))
+    timeout = waits.expectedClientToPoolConnectionTimeout(
+        util.getMaxFailures(len(philCli.nodeReg))
+    )
+    philCli.looper.run(eventually(checkClientConnected,
+                                  philCli.activeClient,
+                                  timeout=timeout))
+
+    timeout = waits.expectedClientToPoolConnectionTimeout(
+        util.getMaxFailures(len(philCli.nodeReg))
+    )
+    poolNodesStarted.looper.run(eventually(checkNodeConnected,
+                                           list(
+                                               poolNodesStarted.nodes.values()),
+                                           timeout=timeout))
+    return newNodeVals
