+++ sovrin_client/test/cli/test_node.py
@@ -1,20 +1,11 @@
 import pytest
-from plenum.common import util
-
-from plenum.test import waits
-from stp_core.loop.eventually import eventually
-
-from stp_core.network.port_dispenser import genHa
 from plenum.common.constants import NODE_IP, CLIENT_IP, CLIENT_PORT, NODE_PORT, \
-    ALIAS, CLIENT_STACK_SUFFIX
+    ALIAS
 from plenum.common.util import randomString
 from plenum.test.cli.helper import exitFromCli
+from stp_core.network.port_dispenser import genHa
 
-
-def doNodeCmd(do, nodeVals, expMsgs=None):
-    expect = expMsgs or ['Node request completed']
-    do('send NODE dest={newNodeIdr} data={newNodeData}',
-       within=8, expect=expect, mapper=nodeVals)
+from sovrin_client.test.cli.helper import doSendNodeCmd
 
 
 @pytest.fixture(scope="module")
@@ -29,51 +20,6 @@ def tconf(tconf, request):
     return tconf
 
 
-@pytest.fixture(scope="module")
-def newNodeAdded(be, do, poolNodesStarted, philCli, newStewardCli,
-                 connectedToTest, newNodeVals):
-    be(philCli)
-
-    if not philCli._isConnectedToAnyEnv():
-        do('connect test', within=3,
-           expect=connectedToTest)
-
-    be(newStewardCli)
-    doNodeCmd(do, newNodeVals)
-    newNodeData = newNodeVals["newNodeData"]
-
-    def checkClientConnected(client):
-        name = newNodeData[ALIAS] + CLIENT_STACK_SUFFIX
-        assert name in client.nodeReg
-
-    def checkNodeConnected(nodes):
-        for node in nodes:
-            name = newNodeData[ALIAS]
-            assert name in node.nodeReg
-
-    timeout = waits.expectedClientToPoolConnectionTimeout(
-        util.getMaxFailures(len(philCli.nodeReg))
-    )
-    newStewardCli.looper.run(eventually(checkClientConnected,
-                                        newStewardCli.activeClient,
-                                        timeout=timeout))
-    timeout = waits.expectedClientToPoolConnectionTimeout(
-        util.getMaxFailures(len(philCli.nodeReg))
-    )
-    philCli.looper.run(eventually(checkClientConnected,
-                                  philCli.activeClient,
-                                  timeout=timeout))
-
-    timeout = waits.expectedClientToPoolConnectionTimeout(
-        util.getMaxFailures(len(philCli.nodeReg))
-    )
-    poolNodesStarted.looper.run(eventually(checkNodeConnected,
-                                           list(
-                                               poolNodesStarted.nodes.values()),
-                                           timeout=timeout))
-    return newNodeVals
-
-
 def testAddNewNode(newNodeAdded):
     pass
 
@@ -81,7 +27,7 @@ def testAddNewNode(newNodeAdded):
 def testConsecutiveAddSameNodeWithoutAnyChange(be, do, newStewardCli,
                                                newNodeVals, newNodeAdded):
     be(newStewardCli)
-    doNodeCmd(do, newNodeVals,
+    doSendNodeCmd(do, newNodeVals,
               expMsgs=['node already has the same data as requested'])
     exitFromCli(do)
 
@@ -95,7 +41,7 @@ def testConsecutiveAddSameNodeWithNodeAndClientPortSame(be, do, newStewardCli,
     newNodeVals['newNodeData'][NODE_PORT] = nodePort
     newNodeVals['newNodeData'][CLIENT_IP] = nodeIp
     newNodeVals['newNodeData'][CLIENT_PORT] = nodePort
-    doNodeCmd(do, newNodeVals,
+    doSendNodeCmd(do, newNodeVals,
               expMsgs=["node and client ha can't be same"])
     exitFromCli(do)
 
@@ -109,7 +55,7 @@ def testConsecutiveAddSameNodeWithNonAliasChange(be, do, newStewardCli,
     newNodeVals['newNodeData'][NODE_PORT] = nodePort
     newNodeVals['newNodeData'][CLIENT_IP] = nodeIp
     newNodeVals['newNodeData'][CLIENT_PORT] = clientPort
-    doNodeCmd(do, newNodeVals)
+    doSendNodeCmd(do, newNodeVals)
     exitFromCli(do)
 
 
@@ -118,6 +64,6 @@ def testConsecutiveAddSameNodeWithOnlyAliasChange(be, do,
                                                   newNodeAdded):
     be(newStewardCli)
     newNodeVals['newNodeData'][ALIAS] = randomString(6)
-    doNodeCmd(do, newNodeVals,
+    doSendNodeCmd(do, newNodeVals,
               expMsgs=['existing data has conflicts with request data'])
     exitFromCli(do)
